requirements.txt
alembic==1.15.2
blinker==1.9.0
boto3==1.38.16
botocore==1.38.16
clapper==1.3.1
click==8.2.0
click-plugins==1.1.1
colorama==0.4.6
Flask==3.1.1
Flask-JWT-Extended==4.7.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==5.1
s3transfer==0.12.0
six==1.17.0
SQLAlchemy==2.0.41
tomli==2.2.1
tomli_w==1.2.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
xdg==6.0.0
